PoC: library call detector
This is a PoC for detecting forbiden call to some functions of a
library. It must be given a file as an argument with the list of
forbiden functions (one function per line).

For now, it must be run with LD_BIND_NOW environement variable set, so
that it can detect the calls.

It only works with the libc for the moment.

Signed-off-by: Paul Semel <semelpaul@gmail.com>